Start integration and rework of BuildBuffer
