+++ Space-Zoologist/Assets/Scripts/Managers/FoodSourceManager.cs
@@ -16,12 +16,10 @@ public class FoodSourceManager : GridObjectManager
     [SerializeField] private GameObject foodSourcePrefab = default;
     public Color constructionColor = new Color(0.5f, 1f, 0.5f, 1f);//Green
     private GridSystem m_gridSystemReference;
-    private BuildBufferManager m_buildBufferManagerReference;
 
     public void Initialize()
     {
         m_gridSystemReference = GameManager.Instance.m_gridSystem;
-        m_buildBufferManagerReference = GameManager.Instance.m_buildBufferManager;
     }
 
     // this used to be part of initialization, but it seems to not be necessary?
@@ -74,11 +72,11 @@ public class FoodSourceManager : GridObjectManager
             pos.y -= 1;
         }
         m_gridSystemReference.AddFood(m_gridSystemReference.WorldToCell(pos), species.Size, newFoodSourceGameObject);
-        m_buildBufferManagerReference.CreateSquareBuffer(new Vector2Int((int)pos.x, (int)pos.y), ttb, species.Size, this.constructionColor);
+        m_gridSystemReference.CreateSquareBuffer(new Vector2Int((int)pos.x, (int)pos.y), ttb, species.Size, this.constructionColor);
         if (ttb > 0)
         {
             foodSource.isUnderConstruction = true;
-            m_buildBufferManagerReference.ConstructionFinishedCallback(() =>
+            m_gridSystemReference.ConstructionFinishedCallback(() =>
             {
                 foodSource.isUnderConstruction = false;
             });
