--- Space-Zoologist/Assets/Scripts/Plot/BuildBufferManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1e3f6e01f8a46fe40ad76f937e51e98d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
