+++ Space-Zoologist/Assets/Scripts/Plot/Grid/GridSystem.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections;
+using System;
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.Tilemaps;
@@ -28,7 +29,6 @@ public class GridSystem : MonoBehaviour
     [SerializeField] private ReservePartitionManager RPM = default;
     [SerializeField] private PopulationManager PopulationManager = default;
     [SerializeField] public Tilemap Tilemap;
-    private BuildBufferManager buildBufferManager;
     #endregion
 
     #region UI
@@ -48,6 +48,11 @@ public class GridSystem : MonoBehaviour
     [SerializeField] private int quickCheckIterations = 6; //Number of tiles to quick check, if can't reach another tile within this many walks, try to generate new body by performing full check
                                                            // Increment by 2 makes a difference. I.E. even numbers, at least 6 to account for any missing tile in 8 surrounding tiles
 
+    [SerializeField] private GameObject bufferGO;
+    private Dictionary<Vector4, List<ConstructionCountdown>> colorTimesToCCs = new Dictionary<Vector4, List<ConstructionCountdown>>();// For serialization
+    private bool[,] isConstructing;
+    public bool IsConstructing(int x, int y) => (x < isConstructing.GetLength(0) && y < isConstructing.GetLength(1)) ? isConstructing[x, y] : false;
+    private Action constructionFinishedCallback = null;
 
     public Vector3Int startTile = default;
     List<Vector3Int> HighlightedTiles;
@@ -64,12 +69,21 @@ public class GridSystem : MonoBehaviour
         this.previewBodies = new List<LiquidBody>();
         this.RemovedTiles = new HashSet<Vector3Int>();
         this.Tilemap.ClearAllTiles();
+
+        LevelData levelData = GameManager.Instance.LevelData;
+        if (levelData == null)
+        {
+            Debug.LogWarning("Level data reference not found, using default width and height values");
+            this.isConstructing = new bool[100, 100];
+        }
+        else
+        {
+            this.isConstructing = new bool[levelData.MapWidth, levelData.MapHeight];
+        }
     }
 
     private void Start()
     {
-        this.buildBufferManager = FindObjectOfType<BuildBufferManager>();
-
         for (int i = 0; i < ReserveWidth; ++i)
         {
             for (int j = 0; j < ReserveHeight; ++j)
@@ -485,6 +499,161 @@ public class GridSystem : MonoBehaviour
         GameManager.Instance.m_playerController.CanUseIngameControls = onOff;
         PauseButton.SetActive(onOff);
         NextDayButton.SetActive(onOff);
+    }
+    #endregion
+
+    #region Building Functions
+    public void CreateUnitBuffer(Vector2Int pos, int time, Color color, int progress = -1)
+    {
+        if (time == 0 || time == -1)
+        {
+            return;
+        }
+        this.isConstructing[pos.x, pos.y] = true;
+        GameObject newGo = Instantiate(this.bufferGO, this.gameObject.transform);
+        //Debug.Log("Placing item under constuction");
+        newGo.transform.position = new Vector3(pos.x + 0.5f, pos.y + 0.5f, 0);
+        color.a = 1; //Enforce alpha channel to be 1, prevent human error       
+        newGo.GetComponent<SpriteRenderer>().color = color;
+        ConstructionCountdown cc = newGo.GetComponent<ConstructionCountdown>();
+        cc.Initialize(pos, time, progress);
+        Vector4 colorTimePair = new Vector4(color.r, color.g, color.b, time);
+        if (!this.colorTimesToCCs.ContainsKey(colorTimePair))
+        {
+            this.colorTimesToCCs.Add(colorTimePair, new List<ConstructionCountdown>());
+        }
+        this.colorTimesToCCs[colorTimePair].Add(cc);
+
+    }
+    public void ConstructionFinishedCallback(Action action)
+    {
+        constructionFinishedCallback += action;
+    }
+    public void CreateSquareBuffer(Vector2Int pos, int time, int size, Color color, int progress = -1)
+    {
+        for (int i = 0; i < size; i++)
+        {
+            for (int j = 0; j < size; j++)
+            {
+                this.CreateUnitBuffer(new Vector2Int(pos.x + i, pos.y + j), time, color, progress);
+            }
+        }
+    }
+
+    public ConstructionCountdown GetBuffer(Vector2Int pos)
+    {
+        if (!this.isConstructing[pos.x, pos.y])
+        {
+            return null;
+        }
+        Vector4[] av4 = this.colorTimesToCCs.Keys.ToArray();
+        List<Vector3Int> changedTiles = new List<Vector3Int>();
+        List<ConstructionCountdown>[] aCCs = this.colorTimesToCCs.Values.ToArray();
+        for (int i = this.colorTimesToCCs.Values.Count - 1; i >= 0; i--)
+        {
+            List<ConstructionCountdown> ccs = aCCs[i];
+            for (int j = ccs.Count - 1; j >= 0; j--)
+            {
+                ConstructionCountdown cc = ccs[j];
+                if (pos == cc.position)
+                {
+                    return cc;
+                }
+            }
+        }
+        return null;
+    }
+    public void DestroyBuffer(Vector2Int pos, int size = 1)
+    {
+        if (!this.isConstructing[pos.x, pos.y])
+        {
+            Debug.Log("No construction buffer to remove, proceeding");
+            return;
+        }
+        HashSet<Vector2Int> bufferToRemove = new HashSet<Vector2Int>();
+        for (int i = 0; i < size; i++)
+        {
+            for (int j = 0; j < size; j++)
+            {
+                bufferToRemove.Add(new Vector2Int(pos.x + i, pos.y + j));
+            }
+
+        }
+        Vector4[] av4 = this.colorTimesToCCs.Keys.ToArray();
+        List<Vector3Int> changedTiles = new List<Vector3Int>();
+        List<ConstructionCountdown>[] aCCs = this.colorTimesToCCs.Values.ToArray();
+        for (int i = this.colorTimesToCCs.Values.Count - 1; i >= 0; i--)
+        {
+            List<ConstructionCountdown> ccs = aCCs[i];
+            for (int j = ccs.Count - 1; j >= 0; j--)
+            {
+                ConstructionCountdown cc = ccs[j];
+                if (bufferToRemove.Contains(cc.position))
+                {
+                    this.isConstructing[(int)cc.position.x, (int)cc.position.y] = false;
+                    ccs.RemoveAt(j);
+                    changedTiles.Add(new Vector3Int(cc.position.x, cc.position.y, 0));
+                    Destroy(cc.gameObject);
+                }
+            }
+            if (ccs.Count == 0)
+            {
+                this.colorTimesToCCs.Remove(av4[i]);
+            }
+        }
+        //Report updates to RPM
+        if (changedTiles.Count > 0)
+        {
+            GameManager.Instance.m_reservePartitionManager.UpdateAccessMapChangedAt(changedTiles);
+        }
+    }
+    public void CountDown()
+    {
+        Vector4[] av4 = this.colorTimesToCCs.Keys.ToArray();
+        List<Vector3Int> changedTiles = new List<Vector3Int>();
+        List<ConstructionCountdown>[] aCCs = this.colorTimesToCCs.Values.ToArray();
+        for (int i = this.colorTimesToCCs.Values.Count - 1; i >= 0; i--)
+        {
+            List<ConstructionCountdown> ccs = aCCs[i];
+            for (int j = ccs.Count - 1; j >= 0; j--)
+            {
+                ConstructionCountdown cc = ccs[j];
+                cc.CountDown();
+                if (cc.IsFinished())
+                {
+                    this.isConstructing[(int)cc.position.x, (int)cc.position.y] = false;
+                    ccs.RemoveAt(j);
+                    Vector3Int pos = new Vector3Int(cc.position.x, cc.position.y, 0);
+                    changedTiles.Add(pos);
+                    Destroy(cc.gameObject);
+
+                    if (constructionFinishedCallback != null)
+                    {
+                        constructionFinishedCallback();
+                    }
+                }
+            }
+            if (ccs.Count == 0)
+            {
+                this.colorTimesToCCs.Remove(av4[i]);
+            }
+        }
+        //Report updates to RPM
+        if (changedTiles.Count > 0)
+        {
+            GameManager.Instance.m_reservePartitionManager.UpdateAccessMapChangedAt(changedTiles);
+        }
+    }
+
+    private Vector3[] GetPositionsAndTimes(List<ConstructionCountdown> cCs)
+    {
+        Vector3[] positions = new Vector3[cCs.Count];
+        for (int i = 0; i < cCs.Count; i++)
+        {
+            Vector2 pos = cCs[i].position;
+            positions[i] = new Vector3(pos.x, pos.y, cCs[i].progress);
+        }
+        return positions;
     }
     #endregion
 
@@ -1073,7 +1242,7 @@ public class GridSystem : MonoBehaviour
             for (int j = 0; j < ReserveWidth; ++j)
             {
                 Vector3Int loopedPosition = new Vector3Int(j, i, 0);
-                if (Vector3Int.Distance(tilePosition, loopedPosition) <= radius)
+                if (Mathf.Abs(loopedPosition.x - tilePosition.x) + Mathf.Abs(loopedPosition.y - tilePosition.y) <= radius)
                     HighlightTile(loopedPosition, color);
             }
         }
@@ -1181,7 +1350,7 @@ public class GridSystem : MonoBehaviour
             return false;
         }
         if (IsOnWall(pos)) return false;
-        if (this.buildBufferManager.IsConstructing(pos.x, pos.y))
+        if (IsConstructing(pos.x, pos.y))
         {
             return false;
         }
